+++ app/controllers/solid_stack_web/cache_controller.rb
@@ -4,6 +4,7 @@ module SolidStackWeb
       @total_entries   = ::SolidCache::Entry.count
       @total_byte_size = ::SolidCache::Entry.sum(:byte_size)
       @size_stats      = CacheSizeStats.new
+      @timeline        = CacheTimeline.new
     end
   end
 end